+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+-   Only url-encode variables in request body (see [#583](https://github.com/frigus02/RESTer/issues/583)).
+
 ## [4.8.1] - 2022-01-23
 
 ### Fixed